rename: util -> globalutils ( 전역 유틸 패키지 )
