+++ server/edusync/src/main/java/com/codestates/edusync/study/studygroup/controller/StudygroupController.java
@@ -6,7 +6,7 @@ import com.codestates.edusync.study.studygroup.dto.StudygroupResponseDto;
 import com.codestates.edusync.study.studygroup.entity.Studygroup;
 import com.codestates.edusync.study.studygroup.mapper.StudygroupMapper;
 import com.codestates.edusync.study.studygroup.service.StudygroupService;
-import com.codestates.edusync.util.UriCreator;
+import com.codestates.edusync.globalutils.UriCreator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpHeaders;
